Refactored shipment export admin:
- Extracted components from the existing components.
- Moved dedicated states and handlers to speicifc component that are not required in the parent.
- Moved interfaces to the types
- Moved table header constants to constants
